# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Meeting_Chat API",
        "description": "Application for free communication",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Application for free communication"
        }
    ],
    "paths": {
        "/interlocutor/{interlocutorCode}": {
            "get": {
                "tags": [
                    "Interlocutors"
                ],
                "summary": "Show interlocutor",
                "description": "Show interlocutor",
                "operationId": "interlocutor",
                "parameters": [
                    {
                        "name": "interlocutorCode",
                        "in": "path",
                        "description": "Interlocutor code",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetInterlocutor"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/videoChat/join/{roomName}/{interlocutorCode}": {
            "get": {
                "tags": [
                    "VideoChat"
                ],
                "summary": "Join the room",
                "description": "Notice of the entrance to the room",
                "parameters": [
                    {
                        "name": "roomName",
                        "in": "path",
                        "description": "Room name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "interlocutorCode",
                        "in": "path",
                        "description": "Interlocutor code",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/videoChat/leave/{roomName}/{interlocutorCode}": {
            "get": {
                "tags": [
                    "VideoChat"
                ],
                "summary": "Leave the room",
                "description": "Notification to leave the room",
                "parameters": [
                    {
                        "name": "roomName",
                        "in": "path",
                        "description": "Room name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "interlocutorCode",
                        "in": "path",
                        "description": "Interlocutor code",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/videoChat/relayIce/{roomName}/{interlocutorCode}": {
            "post": {
                "tags": [
                    "VideoChat"
                ],
                "summary": "Transfer ice candidate",
                "description": "Transfer ice candidate",
                "parameters": [
                    {
                        "name": "roomName",
                        "in": "path",
                        "description": "Room name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "interlocutorCode",
                        "in": "path",
                        "description": "Interlocutor code",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "iceCandidate": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/videoChat/relaySdp/{roomName}/{interlocutorCode}": {
            "post": {
                "tags": [
                    "VideoChat"
                ],
                "summary": "SDP data transfer",
                "description": "SDP data transfer (streams, media data)",
                "parameters": [
                    {
                        "name": "roomName",
                        "in": "path",
                        "description": "Room name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "interlocutorCode",
                        "in": "path",
                        "description": "Interlocutor code",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "sessionDescription": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/videoChat/mute/{roomName}/{interlocutorCode}": {
            "post": {
                "tags": [
                    "VideoChat"
                ],
                "summary": "Notify muting",
                "description": "Notify when audio or video is muted",
                "parameters": [
                    {
                        "name": "roomName",
                        "in": "path",
                        "description": "Room name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "interlocutorCode",
                        "in": "path",
                        "description": "Interlocutor code",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "properties": {
                                            "value": {
                                                "type": "boolean"
                                            },
                                            "key": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object",
                                        "example": {
                                            "value": true,
                                            "key": "audio"
                                        }
                                    },
                                    {
                                        "properties": {
                                            "value": {
                                                "type": "boolean"
                                            },
                                            "key": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object",
                                        "example": {
                                            "value": false,
                                            "key": "video"
                                        }
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "GetInterlocutor": {
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "interlocutorName": {
                        "type": "string"
                    },
                    "code": {
                        "type": "string"
                    },
                    "user": {
                        "oneOf": [
                            {
                                "type": "null"
                            },
                            {
                                "$ref": "#/components/schemas/GetUser"
                            }
                        ]
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "type": "object",
                "example": {
                    "id": 1,
                    "interlocutorName": "Test",
                    "code": "23984y-uioe897-829347g-34879gb23",
                    "user": "null",
                    "created_at": "2023-11-06T15:34:03.000000Z",
                    "updated_at": "2023-11-06T15:34:03.000000Z"
                }
            },
...
            "GetUser": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "gender": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "currentAddress": {
                        "type": "string"
                    },
                    "permanantAddress": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "birthdey": {
                        "type": "string"
                    },
                    "email_verified": {
                        "type": "string"
                    },
                    "avatarPath": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "type": "object",
                "example": {
                    "id": 1,
                    "firstName": "Test",
                    "lastName": "User",
                    "gender": "Female",
                    "phone": "+380501234567",
                    "currentAddress": "Beech Creek, PA, Pennsylvania",
                    "permanantAddress": "Arlington Heights, IL, Illinois",
                    "email": "12345@gmail.com",
                    "birthday": "1975-05-21",
                    "email_verified_at": "2023-11-06T15:34:03.000000Z",
                    "avatarPath": "avatar/127486g-2397y4-28397-2387h4.png",
                    "email_verified": false,
                    "created_at": "2023-11-06T15:34:03.000000Z",
                    "updated_at": "2023-11-06T15:34:03.000000Z"
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}